Set up collection view color and size
